requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
openai==1.81.0
requests==2.31.0
pytest==8.3.5
httpx==0.28.1
python-multipart==0.0.20
loguru==0.7.3
numpy<2.0.0